# services/saep-parent/saep-workforce/src/main/java/com/saep/workforce/model/DigitalProfessional.java
package com.saep.workforce.model;

import jakarta.persistence.*;
import lombok.Getter;
import lombok.Setter;
import java.time.ZonedDateTime;
import java.util.UUID;

@Entity
@Table(name = "workforce_professionals")
@Getter
@Setter
public class DigitalProfessional {

    @Id
    @GeneratedValue(strategy = GenerationType.UUID)
    private UUID id;

    @Column(name = "tenant_id", nullable = false)
    private UUID tenantId;

    @Column(name = "profession_id", nullable = false)
    private UUID professionId;

    @Column(nullable = false)
    private String name;

    @Column(nullable = false)
    private String status = "AVAILABLE";

    @Column(name = "created_at")
    private ZonedDateTime createdAt;

    @Column(name = "updated_at")
    private ZonedDateTime updatedAt;

    @PrePersist
    public void prePersist() {
        this.createdAt = ZonedDateTime.now();
        this.updatedAt = ZonedDateTime.now();
    }

    @PreUpdate
    public void preUpdate() {
        this.updatedAt = ZonedDateTime.now();
    }
}